ByteEncoding
Added the ByteEncoding abstract class, which will serve as the base for all encodings with a code unit size of 1 (i.e. which don't have endianness concerns).

This class provides the IgnoreBOM property, which controls the encoder's behaviour when it encounters a byte order mark in a byte stream.

The UTF8 class now extends this class, rather than extending Encoding directly.