Grapheme Fix
The Grapheme class has three use cases:

1. A thin wrapper around a pointer into a string (i.e. Mode::NonOwning)
2. An owner of multiple code points (i.e. Mode::Owning)
3. An owner of a single code point, which can reside directly inside the Grapheme, eliminating the need for a heap operation (i.e. Mode::Single)

Previously the anonymous union in the Grapheme class only contained a CodePoint and a CodePoint *, which didn't reflect the reality of Mode::NonOwning.  Accordingly the union has been updated to also contain a const CodePoint *, and the code dealing with copying/moving Graphemes has been updated to match.

Also removed the mutable begin and end overloads, as it would be possible for a CodePoint * iterator to be returned from a non-owning Grapheme, resulting in undefined/unintended behaviour.