String Locale Support
Added Locale support to the String class.

Rather than each String method taking a Locale, the String class internally stores a pointer to a Locale, which is used for all operations on that string.  This pointer can be null (as it is by default) in which case the default locale is used.