String Trim Fix
Previously the trim_front and trim_rear private methods of the String class acquired the String's locale internally whenever it was needed.

This presents two problems:

1. A call will be made to Locale::Get for each and every code point, which is potentially expensive
2. Each call to Locale::Get performs an atomic read, which means any atomic read for every code point, which is potentially expensive

And a third if and only if the string is using the default locale (i.e. locale==nullptr):

3. If another thread sets the default locale while another string is trimming, the locale used to trim different parts of the same string may be different, as the locale returned by Locale::Get may differ between calls

Accordingly, trim_front and trim_rear now take a const Locale & as a parameter, and the locale is acquired for them by the Trim method.