Condition Rename
Renamed the class that contains information about conditions from the Unicode data files to ConditionInfo.

Going forward there will need to be a class which encapsulates the actual functionality of such conditions (i.e. checks to see if they hold in a given context), and the name "Condition" suits that role much better.

Also tweaked the documentation for ConditionInfo slightly.