To-/FromEnum
Wrote function template helpers for parsing/serializing enumeration values.

The functions take a reference to an array of std::pair objects which map strings to enumeration values, and either an enumeration (to convert to a string) or a string (to convert to an enumeration value), and return the result of the conversion.