ToInteger
Added a function template which allows a string to be converted to an integer of arbitrary type.

Unlike the built-in C++ library functions, this function template can convert to integers of arbitrary type, not merely a handful of types for which overloads are provided.

Also has a very C++-style interface, unlike the C++ library functions, which have a very C-style interface.