CodePoint/Locale Overhaul
In preparation for a new parser (I was using a C# parser written for a previous project, which was incomplete and inadequate) for the UCD, overhauled codepoint.hpp and locale.hpp.

The CodePointInfo class, relying on new enumerations and helper classes, now more completely encapsulates the UCD.  Some fields (mostly CJK- and Bidi-related) are still missing, but will be added in time (the new parser will be much more extensible).

The Array class template, which wraps a C-style array alongside its length, and provides C++-style abstractions on top of it, is no longer a subclass of Locale, and is used extensively in the new CodePointInfo class, providing a more C++-style interface (no naked points with separate fields giving their size).