UCD Parser
Implemented a UCD parser, which reads the flat text file from UCD.zip, and produces C++ source code output.

Currently the paths to the UCD and the output are hard coded, but this will change (the plan is to accept command line arguments).

Neither the DUCET nor compositions data are read/created by this parser.  Those will be separate parsers which will leverage much of the same machinery.

The parser for SpecialCasing.txt and CaseFolding.txt was updated to fix bugs, and to complete its interface (add output capability).