Parser Output Changes
Had to make several changes to the way the UCD parser generates output:

Changed the way that std::optional and Unicode::Array objects were serialized.  Previously std::optionals were output as simply the value itself (if the std::optional was to be engaged) or as "std::nullopt" (if the optional was to be disengaged).  Unicode::Arrays were output as "{nullptr,0}" if they were to be empty.  GCC would not compile this.  It would balloon up to 6 gigabytes of memory usage, take well over an hour, and eventually fail and die with a segmentation fault/internal compiler error.  Outputting std::optionals as either "{}" (if to be disengaged) or the value enclosed in curly braces (if to be engaged), and outputting empty Unicode::Arrays as "{}" ameliorated this completely and reduced the compile time to less than a minute.

Under the previous scheme the parser output an array of std::pairs, which "mapped" a CodePoint::Type to a CodePointInfo (allowing the array to be searched).  GCC compiled this, but for some reason it caused a segmentation fault in the constructor of some std::pair during startup.  Why the constructor of any std::pairs were being invoked at all given that the memberwise constructor is constexpr (verified in the bits/stl_pair.h header that GCC supplies) remains a mystery.  To fix this the parser now outputs an array of CodePointInfos, the definition of CodePointInfo having been changed to include a CodePoint::Type as its first member.  Worth noting that if all entries from the UCD after and including U+20D3 COMBINING SHORT VERTICAL LINE OVERLAY were removed from the output source file this problem disappeared.