CaseConverter
Added the CaseConverter class, which converts Unicode strings to certain casings, as well as performing case folding.

Can perform conversions based on a specific locale, or default to the current locale.

Can perform either simple or full case mappings/foldings.

The ToTitle method is not yet implemented.

The Fold method is not completely implemented -- in order to handle the edge case of strings containing any character with U+0345 COMBINING GREEK YPOGEGRAMMENI in its canonical decomposition, the Normalizer class will have to be updated/completed.