Condition Fix
A bug in Condition::operator () was causing all conditions to be treated as Final_Sigma.

In Condition::operator () there's a loop over a static const array which contains std::pair objects mapping strings (from the UCD) to function pointers to the implementations of those conditions.  This loop is supposed to continue until a match with the string Condition objects store internally is found, at which point the corresponding function pointer will be invoked.  However the string matching check was erroneously omitted, so this loop unconditionally executed the function pointer corresponding to the first element in the mapping array, which happened to be that of Final_Sigma.