Encoding Changes/Additions
Added the Endianness value to the EncodingErrorType enumeration, to be used when errors related to ambiguous endianness are encountered while decoding.

Changed the documentation for the Encoding::Count method to reflect the fact that it returns the number of code units -- rather than bytes -- that will be required to represent a certain code point.