Formatted Output
Added a templated operator << overload which accepts a std::basic_ostream (templated on any template parameters) and a Unicode string.  This function outputs the string to the output stream-in-question.

The encoding is chosen based on the endianness of the current platform and the size of the character type the output stream is templated on:

- For 1 byte character types, UTF-8 is used
- For 2 byte character types, UTF-16 with the same endianness as the machine is used
- For 4 byte character types, UTF-32 with the same endianness as the machine is used
- For all other character types, a static_assert causes a compile time error