Array Additions
Added a new constructor to the Array class template.  The new constructor is constexpr (like the other constructors) and takes a reference to a C array.  It is templated, deducing the size of the array through a std::size_t template parameter.

Added a global helper function template MakeArray, which deduces the type of a C array, and returns an Array object encapsulating it.

This functionality was previously in getdata.cpp as a private helper function template.  Accordingly modified getdata.cpp to use the global functionality, greatly simplifying the code in that file.