Variable Weights
Updated the DUCET parser and UCD utilities to extract whether a collation element has a variable weight, in addition to the collation element itself.

This necessitated changes to the structures in which this data is stored.

Simultaneously renamed the Collation global function, and the Collation property of the Locale class to CollationTable.  This name better reflects what the function does/property is.