Collator
Implemented the Collator class, which collates Unicode strings.

Lacks documentation.

VariableWeighting::ShiftTrimmed is not implemented yet (it operates identically to VariableWeighting::Shifted -- the final pass to trim 0xFFFF weights is unimplemented).

Getters/setters for options are missing.

Several swathes of functionality are completely untested (such as backwards level 2, and VariableWeighting settings other than NonIgnorable).

Collation element derivation for ideographs is unimplemented, as the CodePointInfo structure does not presently contain the Unified_Ideograph property, which is necessary to implement this functionality.