Collation Elements Retrieval Fix
Fixed a bug in the way that collation elements were retrieved from the collation element table.

When matching the longest initial substring, if the first character could match by itself or at the beginning of a sequence, the search must continue (otherwise contractions will be missed).  However, when this continued search failed, the retriever failed to backtrack to the last match, instead fetching an erroneous contraction.