Blanked Fix
Fixed a bug in the way the Collator handled variable collation elements when the variable weighting setting was Blanked.

The Collator was only checking that the previous collation element was variable before treating the current collation element as ignorable, rather than checking that the previous collation element was variable and that the current collation element is primary ignorable.