Shifted Fix
Fixed a bug in the way the Collator was generating sort keys when the variable ordering was set to Shifted.

When the variable ordering is Shifted or ShiftTrimmed, the Collator should generate sort keys as if the variable ordering is Blanked, except at the quaternary level.  The old, erroneous behaviour didn't reflect this at all, instead generating sort keys as if the variable ordering was NonIgnorable, except at the quaternary level.