Backwards Fix
Fixed a bug that was causing a segmentation fault when the second level was backwards.  Iterators were being passed to MakeReverseIterator in the wrong order (i.e. the begin iterator was being made into the reverse begin, and the end being made into the reverse end, rather than vice versa).