Catch (Unit Testing Framework)
Added a dependency on Catch -- a C++ header-only unit testing framework by philsquared (see updated README.md for more information).

As Catch is a dependency, and not part of this repository, added an entry to ignore its header to .gitignore.

Some of the options I was passing to GCC were breaking Catch, specifically -fno-elide-constructors, which was causing a link time failure.  Accordingly this option is no longer specified in the Makefile.  Also removed a define for DEBUG since that's not currently used in this project, and just happened to be in the Makefile template I reused from another project.

Catch was introducing a #define of IN, which was breaking the LineBreak enumeration.  Added preprocessor guards to fix this.