CaseConverter Fix
SpecialCasing.txt specifies:

IMPORTANT-when iota-subscript (0345) is uppercased or titlecased, the result will be incorrect unless the iota-subscript is moved to the end of any sequence of combining marks. Otherwise, the accents will go on the capital iota.  This process can be achieved by first transforming the text to NFC before casing.  E.g. <alpha><iota_subscript><acute> is uppercased to <ALPHA><acute><IOTA>

The CaseConverter class was not doing this.

In the process of adding the check and conditional normalization before uppercasing, cleaned up some of the CaseConverter code by switching custom written searching loops to use std::find_if, and by changing manual conversions of std::vector iterators to pointer typed iterators to use the templates unicode/vector.hpp provides for exactly this purpose.