Normalizer Fix
When normalizing to NFC and faced with a chain of combining marks which could successively combine with the same starter to create successive composed characters the Normalizer class was generating incorrect output.

For example in:

GREEK SMALL LETTER ALPHA
COMBINING ACUTE ACCENT
COMBINING GREEK YPOGEGRAMMENI

The GREEK SMALL LETTER ALPHA and COMBINING ACUTE ACCENT compose to form GREEK SMALL LETTER ALPHA WITH TONOS.  This then composes with COMBINING GREEK YPOGEGRAMMENI to form GREEK SMALL LETTER ALPHA WITH OXIA AND YPOGEGRAMMENI.

In this instance the Normalizer class was erroneously generating the output:

GREEK SMALL LETTER ALPHA WITH YPOGEGRAMMENI

As the Normalizer loops over a string trying to find compositions, it is managing two sequences of code points:

1. The input sequence
2. The output sequence

As an optimization, the Normalizer only creates a new std::vector of CodePoints if a composition must be made.  A large number of input strings do not require composition, and therefore for a large number of strings the creation of a new std::vector can be avoided, increasing speed and decreasing memory requirements.

However, after making a composition and creating a new std::vector of CodePoints including the composition, the Normalizer was still referring to starters in the old std::vector when looking up compositions.  This had no effect when many different starters were variously replaced with compositions, since each starter was only composed once, however when many combining marks attached to the same starter composed, this had the effect of causing the earlier compositions to be discarded.

For example: In the example given above, the Normalizer would see GREEK SMALL LETTER ALPHA and COMBINING ACUTE ACCENT, and combine them into GREEK SMALL LETTER ALPHA WITH TONOS.  Then the Normalizer would see COMBINING GREEK YPOGEGRAMMENI, but would look at the starter in the input sequence, which would be GREEK SMALL LETTER ALPHA.  It would then overwrite the GREEK SMALL LETTER ALPHA WITH TONOS with GREEK SMALL LETTER ALPHA WITH YPOGEGRAMMENI, acting as if the COMBINING ACUTE ACCENT had never been in the input sequence at all.