Unit Test Tweaks
Added a toString method for the String class to the unit tester.  This allows Catch to properly output String objects when breaking out conditions for display.

For whatever reason (I don't know the internals of Catch) this had to be placed before Catch was included, which drove the include directive for Catch below the other include directives, which eliminated the need for preprocessor hacks to protect LineBreak::IN from unwanted macro substitution.