RegexLiteral
Implemented a skeleton for the RegexLiteral class.  During parsing/compilation, if a character/sequence does not correspond to any pattern element, it is compiled as a literal, meaning that it will match the exact sequence-in-question.

Also added the ToString method to the RegexPatternElement abstract class, which will be used for debugging, obtaining a descriptive string representation of a pattern element.  The class the ToString method will return instances of -- RegexToString -- was also added.