Converter Reorganization
Reorganized the way the Converter class is implemented.

Previously the Converter class was implemented directly, exposing the fact that it had two template parameters (one of which was a dummy for SFINAE).

The reorganization renames the Converter class to ConverterImpl, and implements a Converter class which simply inherits from this ConverterImpl class, hiding the SFINAE template parameter.