String Conversion Improvements
Changed the CString class -- which encapsulates a C-style string, using RAII to manage its dynamically allocated memory, and being implicitly convertible to const char * -- to be a template.  Now it may encapsulate a C-style string which uses characters of any type.

Rewrote the String::ToCString method to take a template parameter (defaulted to char) to determine what type of characters to use for the resulting C-style string.  An encoding is chosen appropriately based on the size of the requested characters.

Added the String::ToOSString method, which returns a CString using a character type suitable to be passed to underlying operating system APIs.  For Windows this is wchar_t (UTF-16LE), for other systems this is char (UTF-8).

Added the String::ToString template method.  It returns a std::basic_string using a character type given by its template parameter (defaulted to char).