String Converter Integration
Added the IsConvertible helper template class to converter.hpp.  This template can be used to determine if a certain type is convertible to/from String via the Converter class.

Added constructor/assignment overloads to the String class such that for an object of any type T such that IsConvertible<T>::Value==true, that object may be used to construct a String, or may be assigned to a String, and this will result in the String containing the string representation of that object as obtained by the Converter template class.