UTF8 Tweak/Fix
Previously when the UTF-8 decoder encountered a non-continuation byte in a multi-byte sequence, it would consume that non-continuation byte and report an error.  This could lead to perfectly acceptable code units being removed from the stream because they were preceded by a malformed sequence.  Corrected this so that if a code unit is found not to be a continuation byte, the stream is rewound before returning.