Encoding Fix
Fixed an issue in the way Encoding::Decode handled EncodingErrorType::UnexpectedEnd errors.

The correct behaviour is to end the decoding immediately, after taking whatever action was specified.  However, this case was not detected correctly, and accordingly decoding would continue after the error (unless the action for unexpected end was to throw), resulting -- in some cases -- in EncodingErrorType::Strict being raised.