UTF16 Fix
After encountering a lead surrogate which is not followed by a trail surrogate, the UTF-16 decoder backs off, so that if the error action for strict errors is not to throw, the following code unit will be processed rather than being skipped.

However this back off was by one byte, when a UTF-16 code unit is 2 bytes.