Type Changes
Changed the type of CodePoint::Type from std::uint32_t to char32_t to be consistent with the minimal way in which C++11 handles Unicode.

Changed the type of UTF16::CodeUnit from std::uint16_t to char16_t for the same reason.

Changed the constructors and assignment operator overloads of CodePoint so that rather than accepting chars and CodePoint::Types, any integral type is accepted (via template/SFINAE) and simply casted to CodePoint::Type before being assigned to the CodePoint::cp private property.