Input Streams
Added overloads for operator >> for std::basic_istream objects for both String and CodePoint objects.  The overload for Strings extracts one line at a time, the overload for CodePoints extracts one code point at a time.  Neither overload ignores white space.  The overload for CodePoint will extract characters which cause line breaks, the overload for Strings will not.

As with all other utilities that interoperate with the C++ standard library, the encoding to use is chosen based on the width of the character type of the input stream.