RegexCompiler Fix
Fixed a bug caused by undefined behaviour in the RegexCompiler class:

The global, static vector of parsers was being accessed before it was constructed, due to the fact that global constructors are called in an undefined order between files.

Fixed this by acquiring said vector using the Meyer's singleton pattern.