String Conversion
Added the To and Convert method templates to the String class.

The former uses the Converter class template to attempt to convert a string to a certain type, returning a std::optional with the results (disengaged if the conversion failed).

The latter uses the Converter class template to attempt to convert a string to a certain type, throwing on failure.