RegexPatternElementState
Removed the RegexPatternElementState class.

The idea was not mature and did not make sense in the incarnation in which it currently existed.

The concept of a base class which provides a mechanism for pattern elements to store data specific to them is a good one, however, it is necessary for there to be some per-pattern element structure which stores certain common data (such as backtracking information), in addition to an abstract base class for purely private data.

This warrants further consideration/a different implementation/realization.