Regular Expression Compiler Tweaks/Refactor
Added the RegexCompilerState class, which carries around the regular expression compiler's internal state, as well as having convenient methods for manipulating that state, similar to RegexState.

Adjusted the RegexCompiler, RegexParser, and RegexLiteral classes to use this state, rather than specific arguments -- which was limiting and would not have provided flexibility going forward when new options/flags/et cetera proved to be necessary.