RegexCompilerState Change
Due to the fact that many Unicode operations are context-sensitive, it is often necessary to be able to look backward as well as forward within a string.

Accordingly, changed the RegexCompilerState class so that it holds not just a begin (i.e. current position) and end iterator, but a begin, current, and end iterator, which allows parsers to look forward and backward in the pattern for context-sensitive Unicode operations.