Regular Expression Compiler Changes
Modified the RegexCompiler class so that its operator () overload -- which compiles a pattern -- doesn't take individual arguments, but rather an entire RegexCompilerState.

Added a virtual Done method to the RegexCompilerState class, which returns true if the current point in the underlying pattern is where the RegexCompiler should stop compiling.

This will enable an important use case -- compiling only part of a pattern.  This will be necessary for implementing pattern elements which logically contain many other pattern elements -- grouping constructs for example.