RegexCompiler Overhaul
Removed the RegexCompilerState class and merged it into the RegexCompiler class.

This gives the RegexCompiler class some purpose in being a class, rather than a collection of functions, as previously it carried no state information.

Also this provides better encapsulation, as there's no concerns as to how the RegexCompiler is going to communicate with the RegexCompilerState versus how the RegexParser is going to communicate with it.