RegexState Improvements
Improved the interface/implementation of the RegexState class:

RegexState's constructor now takes a const RegexEngine reference, rather than taking a RegexIterator.  This simplifies the interface/invocation and allows more information from the RegexEngine to be harvested into the RegexState during construction in the future, if required.

RegexState::Rewind now rewinds the nested RegexPrivateState -- if any -- before rewinding the RegexEngine.  This is more intuitive, as the RegexEngine will be in the state in which the creator of the RegexPrivateState left it when RegexPrivateState::Rewind is invoked.