RegexMatch Support
Added RegexMatch support to the RegexEngine class.

RegexEngines now carry a reference to a RegexMatch, which RegexPatternElements can then add to/modify.

Also added a new constructor to the RegexEngine class, which takes another RegexEngine and a collection of RegexPatternElements, the newly constructed RegexEngine then has the same logical state as the RegexEngine it was constructed from, but it matches a different pattern.

This greatly simplifies the construction of RegexEngines to match subpatterns.

Also added a Match method (and one overload) to the Regex class, which actually performs a regular expression match.