RegexCompiler Copy Constructor
Added a faux copy constructor to the RegexCompiler class.

This constructor does not actually copy the RegexCompiler-in-question -- due to the nature of std::unique_ptr, RegexCompilers are not copyable -- but rather copies all the information the compiler contains, except the pattern it's compiled and related state.

This greatly simplifies the creation of compilers to compile subpatterns, as in regexcharacterclass.cpp, where the creation of a compiler to compile character sets was greatly simplified.