RegexCharacterClass Fix
Fixed a bug in the RegexCharacterClass class which was causing backtracking to loop infinitely.

When backtracking, the RegexCharacterClass should select the pattern element after the pattern element which matched previously, and start attempting to match from there.  However it was instead selecting the pattern element which matched previously, causing an infinite loop as it formed the same match over and over again.