RegexEngine Current Location Fix/Improvement
Due to the fact that a match may begin anywhere within a string (if a match fails at the first location in the string, the match will be attempted again at the next location, et cetera) it is important due to the context-sensitive nature of some Unicode algorithms (and potential regular expression pattern elements, such as look around) that the context be preserved.

Accordingly the RegexEngine constructor no longer simply takes a begin and end iterator, rather taking a begin, location, and end iterator, meaning that the entire string will always be available for context-sensitive algorithms/pattern elements, but that the match may begin at an arbitrary location within that string.