Regex::Match Overhaul
Added a new Regex::Match overload which allows for streaming matching -- i.e. it takes certain iterators as out parameters, and updates them as the matcher works through the string, allowing successive matches to be obtained by successively calling the method.  The other Regex::Match overloads are now implemented in terms of this overload.

The Regex::Match overload which takes a begin and end iterator now accepts an optional locale parameter, which is used for normalization.

The Regex::Match overload which takes a string now uses the string's locale.