RegexCompiler Capturing Group Support
Added capturing group support to the RegexCompiler class.

The RegexCompiler can now track all RegexPatternElements associated with a certain named or numbered capturing group, as well as automatically generating monotonically increasing capturing group numbers (beginning with 1).

Removed one of the RegexCompiler constructor overloads, as it was obsolete and maintaining it was becoming a nuisance.