RegexAtomic Fix
Fixed a bug which was causing RegexAtomic pattern elements not to appropriately advance the state of the RegexEngine.

RegexAtomic pattern elements use a RegexEngine internally to match against the pattern elements they maintain.  Accordingly the main RegexEngine (which the pattern element is passed as a parameter) must be fast forwarded after the RegexAtomic's internal RegexEngine completes, however this was not done.