RegexEngine::Rewind
Added the Rewind method to the RegexEngine class.  This rewinds each pattern element the RegexEngine has matched.

This is necessary for pattern elements which use a RegexEngine internally, can't backtrack, but may be backtracked over, as when they are backtracked over they must give contained pattern elements the opportunity to clean up.