Regex Fix/Consistency Update
Brought the order of parameters to Regex::Match in line with the loc, begin, end convention used by the rest of the library.

Fixed a bug which was preventing the regular expression engine from forming matches at the end (in left-to-right mode) and beginning (in right-to-left mode) of the input string.