Regular Expression Execution Overhaul
Overhauled the Regex methods used to execute/evaluate regular expressions.

Previously a Regex::Match overload that took a pointer reference was used, however this model was insufficient and buggy -- under certain situations certain matches were not performed and further unwieldy out parameters would be required for it to maintain all the required state information.

Moreover, no facility was directly/conveniently supplied for lazy evaluation of the matches of a regular expression, i.e. for evaluating several -- but not all -- matches within a string.

Reimplemented the matching functionality within the Regex::Iterator class.  This is a ForwardIterator which lazily evaluates each match of a regular expression found in some string.

Added the Regex::Iterable class.  This class wraps a regular expression/string combination, providing begin and end methods which return Regex::Iterators and is a convenient target for range-based for loops.

Added the Regex::Iterate method overloads.  These are helpers which wrap the current regular expression and a string into a Regex::Iterable and return it.  This allows a Regex::Iterable to be easily obtained as the target of a range-based for loop.

The Regex::Match and Regex::Matches methods and overloads are now implemented in terms of range-based for loops and the Regex::Iterate method.