Regular Expression Compiler Group-Handling Improvements
There is now a method -- RegexCompiler::GetLastCaptureNumber -- for obtaining the last automatically generated capture number without causing that number to be incremented.

Modified the internal structures the RegexCompiler uses to store numbered and named capturing groups to accomodate the idea of pending references to a capturing group.

Certain regular expression constructs allow for arbitrary groups within a pattern to be recursed.  For this to be implemented, some kind of reference/pointer to the pattern element which implements that group must be obtained.  This is straightforward if such a pattern element has already been compiled, but if one has yet to be compiled matters are complicated.

Accordingly, the RegexCompiler now maintains a list both of pattern elements which capture a certain number/name, and of pointers which are waiting for the address of such a pattern element.  If the address of a pattern element capturing a certain number/name is requested and such a pattern element is present, the pointer-in-question is populated immediately.  Otherwise it is added to a list of pending pointers, and is populated when the compiler finishes.

If, at completion, there are any pending pointers which cannot be satisfied, an exception is thrown.

This functionality is exposed through the RegexCompiler::GetCapturingGroup overloads.