Regex::Iterable Fix
Fixed an issue which was causing all strings which required normalization to NFD to be treated as the empty string.

Regex::Iterable maintains two things internally:

1. A std::optional<String> which holds a string if the input string had to be normalized
2. Begin and end iterators in the form of const CodePoint * (as elsewhere throughout the library)

When an input string requires normalization, the std::optional<String> is engaged and populated with the normalized string.  The begin and end iterators should then be set to the begin and end iterators of that string, but they were instead both set to nullptr, causing the regular expression engine to attempt to match against the empty string (since begin==end, therefore the string is empty).