Backtracking Fixes
Fixed several problems with the way that backtracking was implemented, and the pattern elements that supported backtracking.

During backtracking, the RegexState::Rewind method was being invoked for each element before it was determined whether or not backtracking ended with that element.  This meant that when an element was invoked after backtracking, the engine had been rewound to the state it was in before the element was first invoked.

This didn't reflect the purpose/implementation of backtracking.  If the engine has been rewound to the state it was in before the pattern element matched, how does it make sense for the pattern element to match based on the state established from the preceding match?  The regular expression engine -- by calling Rewind -- expressed that it wished for that previous match not to have happened.  This logical flaw led to implementation complexity in derived class overrides of RegexPrivateState::Rewind.  A lot of logic and checking went into determining whether it was actually a rewind -- i.e. whether the internal state of the RegexPrivateState should be rewound -- or not.

Accordingly RegexEngines now invoke RegexState::Rewind only immediately before they delete a RegexState and move to the preceding pattern element.  Derived class overrides of RegexPrivateState::Rewind have been updated to reflect this, removing the complex logic present previously.

Moreover, some pattern elements were not properly updating the flags on their RegexState when called after backtracking.  Prevalently, the RegexState was not updated after a failure.  This caused the regular expression engine to loop infinitely, as the engine saw that the current element was flagged as being able to backtrack (from previous, successful matches), invoked it, and the element's internal logic dictated that it could not backtrack, causing it to fail, causing backtracking, ad infinitum.