RegexCompiler::Pop Problem/Inconsistency
The RegexCompiler class provides facilities to compile RegexPatternElements which are formed incrementally.  RegexLiteral is an example of such a pattern element: A contiguous string of literals -- i.e. code points/sequences which the regular expression engine should not interpret specially -- are aggregated into a single RegexPatternElement.  This has the advantage of reducing:

1. The number of virtual calls during pattern matching/execution
2. The number of heap allocations/deallocations which must be performed during compilation/teardown
3. The number of RegexState objects which must be pushed into a vector/backtracked over during matching/execution

However, the RegexCompiler::Pop method acted as though this mechanism did not exist.  This meant, for example, that when repetition constructs were added, the following regular expressions would've been treated as equivalent:

- (?:aaaa){4}
- aaaa{4}

Which is clearly bizarre, erroneous, and unintended.

The solution is to allow RegexParsers the opportunity to split RegexPatternElements up in response to a call to RegexCompiler::Pop.  This is accomplished through a new method on the RegexParser class -- GetLast -- which takes a reference to a RegexPatternElement, and returns a RegexCompiler::Element -- the result of abstracting the last logical RegexPatternElement.