Integer Conversion Fix
Due to the fact that the UCD does not assign integer values to letters which represent hexadecimal values, converting from a hexadecimal string to an integer was failing improperly.

The solution was to -- in addition to checking the UCD -- probe the Locale::Digits array for a digit which represents the same code point under simple case folding, and then take its associated numerical value.

Added a unit test to verify that this fix is working correctly.