Range Fix
The range character set construct (i.e. [a-z]) was previously taking a single code point on either side and using it as the begin and end of the range.  While this satisfies some of the requirements of the range construct (e.g. it will correctly parse and compile [a-z]) it does not satisfy others (e.g. it will incorrectly parse and compile [\x00-\x29]).

Moved the range code from regexrange.cpp (which has been deleted) to regexliteral.cpp so the code for parsing escapes may be shared.

Also cleaned up the parsing code a bit.